Feature/add unit tests

+++ src/service/item-service.ts
@@ -2,15 +2,18 @@ import { Service } from 'typedi';
 import { InjectRepository } from 'typeorm-typedi-extensions';
 import { IItem } from '../model/i-item';
 import { ItemRepository } from '../repository/item-repository';
-import Logger from "../lib/logger";
+import logger from "../config/logger";
 import { DeleteResult, UpdateResult } from 'typeorm';
+import { WishlistRepository } from '../repository/wishlist-repository';
 
 @Service()
 export class ItemService {
 
   constructor(
     @InjectRepository()
-    private readonly itemRepository: ItemRepository
+    private readonly itemRepository: ItemRepository,
+    @InjectRepository()
+    private readonly wishlistRepository: WishlistRepository
   ) {
   }
 
@@ -26,13 +29,29 @@ export class ItemService {
     });
   }
 
-  public save(itemOptions: IItem): Promise<IItem> {
+  public async save(itemOptions: IItem): Promise<IItem> {
+    const wishlistId = itemOptions.wishlist?.id;
+    let foundWishlist;
+    if (wishlistId) {
+      foundWishlist = await this.wishlistRepository.findOne({
+        id: wishlistId,
+      });
+    }
+
+    // An item must have a wishlist
+    // Because each wishlist must be of an user and all items/wishlists are contexts of an user
+    if (!foundWishlist) {
+      throw new Error('Wishlist not found in the database! An item must have a wishlist as parent.')
+    }
+
     try {
-      return this.itemRepository.save(itemOptions);
+      return this.itemRepository.save({
+        ...itemOptions,
+        wishlist: foundWishlist,
+      });
     } catch (err) {
-      Logger.error(`Error while saving item: ${err}`)
+      logger.error(`Error while saving item: ${err}`)
       throw new Error(`Error while saving item: ${err}`)
-      // todo: catch it in controller | throw new Error(`Error while saving item: ${err}`);
     }
   }
 
@@ -45,13 +64,16 @@ export class ItemService {
         }
       );
     } catch (err) {
-      Logger.error(`Error while updating item: ${err}`)
+      logger.error(`Error while updating item: ${err}`)
       throw new Error(`Error while updating item: ${err}`)
-      // todo: catch it in controller | throw new Error(`Error while updating item: ${err}`);
     }
   }
 
   public async delete(itemId: number): Promise<DeleteResult> {
+    const foundItem = await this.itemRepository.findOne(itemId);
+    if (!foundItem) {
+      throw new Error(`Item with id ${itemId} was not found!`);
+    }
     return this.itemRepository.delete(itemId);
   }
 }
